Sweeping updates to allow for compiliation
- Accounting for update to function signatures for swapchain and time-dependent update
- Updating Cmakelists so that including vpr elements is actually possible
- Update source code with vpr includes so that the files can be included
- Removing unused libraries that have been left in the Cmakelists in some locations

+++ plugin_manager/include/CoreAPIs.hpp
@@ -19,7 +19,7 @@ struct Plugin_API {
     // Fixed-timestep update of logical components not requiring updates w/ timestep value
     void (*LogicalUpdate)(void);
     // Called per frame, dt is frametime. use for time-dependent updates like physical systems and simulations
-    void (*TimeDependentUpdate)(float dt);
+    void (*TimeDependentUpdate)(double dt);
     // For potential future API expansion
     void* ReservedFns[8];
 };
